Fix old and really annoying bug
When a chapter was selected for the first time (either from the list or
from the download widget) even though the right chapter was selected and
marked as read, the first chapter was always selected in the chapters
combo box. The bug didn't occur again after that, only the first time.
The bug was caused by setting the focus to the page label, mandatory
step, otherwise the mangas combo box had the focus and when the user
wanted to scroll down the page, the manga selected was updated. I just
forced the chapters combo box to have the right current chapter,
checking if the chapter names were equaled. I still don't get why this
bug occured in the first place. Some kind of weird stuff involving
asynchronous programmation that I still haven't got, most likely. Also,
remove the "Chapter" first item in chapters combo box, for it's
completely useless.